app/requirements.txt
python-dotenv
flask
flask-cors
cerberus
pandas 
